+++ Packages/com.unity.inputsystem/InputSystem/Devices/InputDeviceBuilder.cs
@@ -61,8 +61,7 @@ namespace UnityEngine.Experimental.Input.Layouts
         {
             if (existingDevice != null && existingDevice.m_DeviceIndex != InputDevice.kInvalidDeviceIndex)
                 throw new InvalidOperationException(
-                    string.Format("Cannot modify control setup of existing device {0} while added to system.",
-                        existingDevice));
+                    $"Cannot modify control setup of existing device {existingDevice} while added to system.");
 
             InstantiateLayout(layout, variants, new InternedString(), null, existingDevice);
             FinalizeControlHierarchy();
@@ -101,7 +100,7 @@ namespace UnityEngine.Experimental.Input.Layouts
                 return match;
 
             if (ReferenceEquals(parent, m_Device))
-                return InputControlPath.TryFindControl(m_Device, string.Format("{0}/{1}", m_Device.name, path));
+                return InputControlPath.TryFindControl(m_Device, $"{m_Device.name}/{path}");
 
             return null;
         }
@@ -138,11 +137,9 @@ namespace UnityEngine.Experimental.Input.Layouts
         {
             var control = GetControl(parent, path);
 
-            var controlOfType = control as TControl;
-            if (controlOfType == null)
-                throw new Exception(string.Format(
-                    "Expected control '{0}' to be of type '{1}' but is of type '{2}' instead!", path,
-                    typeof(TControl).Name, control.GetType().Name));
+            if (!(control is TControl controlOfType))
+                throw new Exception(
+                    $"Expected control '{path}' to be of type '{typeof(TControl).Name}' but is of type '{control.GetType().Name}' instead!");
 
             return controlOfType;
         }
@@ -151,7 +148,7 @@ namespace UnityEngine.Experimental.Input.Layouts
         {
             var control = TryGetControl(path);
             if (control == null)
-                throw new Exception(string.Format("Cannot find input control '{0}'", path));
+                throw new Exception($"Cannot find input control '{path}'");
             return control;
         }
 
@@ -165,7 +162,7 @@ namespace UnityEngine.Experimental.Input.Layouts
         {
             var control = TryGetControl<TControl>(path);
             if (control == null)
-                throw new Exception(string.Format("Cannot find input control '{0}'", path));
+                throw new Exception($"Cannot find input control '{path}'");
             return control;
         }
 
@@ -178,9 +175,8 @@ namespace UnityEngine.Experimental.Input.Layouts
 
             var controlOfType = control as TControl;
             if (controlOfType == null)
-                throw new Exception(string.Format(
-                    "Expected control '{0}' to be of type '{1}' but is of type '{2}' instead!", path,
-                    typeof(TControl).Name, control.GetType().Name));
+                throw new Exception(
+                    $"Expected control '{path}' to be of type '{typeof(TControl).Name}' but is of type '{control.GetType().Name}' instead!");
 
             return controlOfType;
         }
@@ -240,20 +236,18 @@ namespace UnityEngine.Experimental.Input.Layouts
                 control = controlObject as InputControl;
                 if (control == null)
                 {
-                    throw new Exception(string.Format("Type '{0}' referenced by layout '{1}' is not an InputControl",
-                        layout.type.Name, layout.name));
+                    throw new Exception(
+                        $"Type '{layout.type.Name}' referenced by layout '{layout.name}' is not an InputControl");
                 }
             }
 
             // If it's a device, perform some extra work specific to the control
             // hierarchy root.
-            var controlAsDevice = control as InputDevice;
-            if (controlAsDevice != null)
+            if (control is InputDevice controlAsDevice)
             {
                 if (parent != null)
-                    throw new Exception(string.Format(
-                        "Cannot instantiate device layout '{0}' as child of '{1}'; devices must be added at root",
-                        layout.name, parent.path));
+                    throw new Exception(
+                        $"Cannot instantiate device layout '{layout.name}' as child of '{parent.path}'; devices must be added at root");
 
                 m_Device = controlAsDevice;
                 m_Device.m_StateBlock.byteOffset = 0;
@@ -285,9 +279,7 @@ namespace UnityEngine.Experimental.Input.Layouts
                 // Someone did "new InputDeviceBuilder(...)" with a control layout.
                 // We don't support creating control hierarchies without a device at the root.
                 throw new InvalidOperationException(
-                    string.Format(
-                        "Toplevel layout used with InputDeviceBuilder must be a device layout; '{0}' is a control layout",
-                        layout.name));
+                    $"Toplevel layout used with InputDeviceBuilder must be a device layout; '{layout.name}' is a control layout");
             }
 
             // Name defaults to name of layout.
@@ -327,7 +319,7 @@ namespace UnityEngine.Experimental.Input.Layouts
                 // now be blank) but still want crawling down the hierarchy to preserve existing
                 // controls where possible.
                 AddChildControls(layout, variants, control,
-                    existingControl != null ? existingControl.m_ChildrenReadOnly : (ReadOnlyArray<InputControl>?)null,
+                    existingControl?.m_ChildrenReadOnly,
                     ref haveChildrenUsingStateFromOtherControl);
             }
             catch
@@ -355,9 +347,7 @@ namespace UnityEngine.Experimental.Input.Layouts
                     var referencedControl = TryGetControl(control, controlLayout.useStateFrom);
                     if (referencedControl == null)
                         throw new Exception(
-                            string.Format(
-                                "Cannot find control '{0}' referenced in 'useStateFrom' of control '{1}' in layout '{2}'",
-                                controlLayout.useStateFrom, controlLayout.name, layout.name));
+                            $"Cannot find control '{controlLayout.useStateFrom}' referenced in 'useStateFrom' of control '{controlLayout.name}' in layout '{layout.name}'");
 
                     // Copy its state settings.
                     child.m_StateBlock = referencedControl.m_StateBlock;
@@ -396,9 +386,8 @@ namespace UnityEngine.Experimental.Input.Layouts
                 if (controlLayouts[i].isModifyingChildControlByPath)
                 {
                     if (controlLayouts[i].isArray)
-                        throw new NotSupportedException(string.Format(
-                            "Control '{0}' in layout '{1}' is modifying the child of another control but is marked as an array",
-                            controlLayouts[i].name, layout.name));
+                        throw new NotSupportedException(
+                            $"Control '{controlLayouts[i].name}' in layout '{layout.name}' is modifying the child of another control but is marked as an array");
 
                     haveControlLayoutWithPath = true;
                     InsertChildControlOverrides(parent, ref controlLayouts[i]);
@@ -504,18 +493,16 @@ namespace UnityEngine.Experimental.Input.Layouts
 
             ////REVIEW: can we check this in InputControlLayout instead?
             if (string.IsNullOrEmpty(controlItem.layout))
-                throw new Exception(string.Format("Layout has not been set on control '{0}' in '{1}'",
-                    controlItem.name, layout.name));
+                throw new Exception($"Layout has not been set on control '{controlItem.name}' in '{layout.name}'");
 
             // See if there is an override for the control.
             InputControlLayout.ControlItem? controlOverride = null;
             if (m_ChildControlOverrides != null)
             {
-                var path = string.Format("{0}/{1}", parent.path, name);
+                var path = $"{parent.path}/{name}";
                 var pathLowerCase = path.ToLower();
 
-                InputControlLayout.ControlItem match;
-                if (m_ChildControlOverrides.TryGetValue(pathLowerCase, out match))
+                if (m_ChildControlOverrides.TryGetValue(pathLowerCase, out var match))
                     controlOverride = match;
             }
 
@@ -551,8 +538,7 @@ namespace UnityEngine.Experimental.Input.Layouts
             {
                 // Throw better exception that gives more info.
                 throw new Exception(
-                    string.Format("Cannot find layout '{0}' used in control '{1}' of layout '{2}'",
-                        exception.layout, name, layout.name),
+                    $"Cannot find layout '{exception.layout}' used in control '{name}' of layout '{layout.name}'",
                     exception);
             }
 
@@ -608,7 +594,7 @@ namespace UnityEngine.Experimental.Input.Layouts
             ////        of successive re-allocations
 
             // Add usages.
-            var usages = controlOverride != null ? controlOverride.Value.usages : controlItem.usages;
+            var usages = controlOverride?.usages ?? controlItem.usages;
             if (usages.Count > 0)
             {
                 var usageCount = usages.Count;
@@ -652,8 +638,7 @@ namespace UnityEngine.Experimental.Input.Layouts
             var pathLowerCase = path.ToLower();
 
             // See if there are existing overrides for the control.
-            InputControlLayout.ControlItem existingOverrides;
-            if (!m_ChildControlOverrides.TryGetValue(pathLowerCase, out existingOverrides))
+            if (!m_ChildControlOverrides.TryGetValue(pathLowerCase, out var existingOverrides))
             {
                 // So, so just insert our overrides and we're done.
                 m_ChildControlOverrides[pathLowerCase] = controlItem;
@@ -760,14 +745,12 @@ namespace UnityEngine.Experimental.Input.Layouts
             var immediateParent = InputControlPath.TryFindChild(parent, immediateParentPath);
             if (immediateParent == null)
                 throw new Exception(
-                    string.Format("Cannot find parent '{0}' of control '{1}' in layout '{2}'", immediateParentPath,
-                        controlItem.name, layout.name));
+                    $"Cannot find parent '{immediateParentPath}' of control '{controlItem.name}' in layout '{layout.name}'");
 
             var controlName = path.Substring(indexOfSlash + 1);
             if (controlName.Length == 0)
                 throw new Exception(
-                    string.Format("Path cannot end in '/' (control '{0}' in layout '{1}')", controlItem.name,
-                        layout.name));
+                    $"Path cannot end in '/' (control '{controlItem.name}' in layout '{layout.name}')");
 
             // Make room in the device's child array.
             var childStartIndex = immediateParent.m_ChildrenReadOnly.m_StartIndex;
@@ -887,8 +870,8 @@ namespace UnityEngine.Experimental.Input.Layouts
                 var field = objectType.GetField(parameter.name,
                     BindingFlags.IgnoreCase | BindingFlags.Instance | BindingFlags.Public | BindingFlags.NonPublic);
                 if (field == null)
-                    throw new Exception(string.Format("Cannot find public field {0} in {1} (referenced by parameter)",
-                        parameter.name, objectType.Name));
+                    throw new Exception(
+                        $"Cannot find public field {parameter.name} in {objectType.Name} (referenced by parameter)");
 
                 ////REVIEW: can we do this without boxing?
 
@@ -947,9 +930,7 @@ namespace UnityEngine.Experimental.Input.Layouts
             if (control.m_StateBlock.sizeInBits == 0 && children.Count == 0)
             {
                 throw new Exception(
-                    string.Format(
-                        "Control '{0}' with layout '{1}' has no size set and has no children to compute size from",
-                        control.path, control.layout));
+                    $"Control '{control.path}' with layout '{control.layout}' has no size set and has no children to compute size from");
             }
 
             // If there's no children, our job is done.
@@ -1067,7 +1048,7 @@ namespace UnityEngine.Experimental.Input.Layouts
             control.m_StateBlock.sizeInBits = totalSizeInBytes * 8;
         }
 
-        // Finalize array references in the control hierarchy and make all state offets relative to the
+        // Finalize array references in the control hierarchy and make all state offsets relative to the
         // device root.
         private void FinalizeControlHierarchy()
         {