+++ Packages/com.unity.inputsystem/Tests/InputSystem/CoreTests_Layouts.cs
@@ -1436,6 +1436,34 @@ partial class CoreTests
 
     [Test]
     [Category("Layouts")]
+    public void Layouts_NoisyControls_AutomaticallyMakeAllTheirChildrenNoisy()
+    {
+        const string json = @"
+            {
+                ""name"" : ""TestLayout"",
+                ""controls"" : [
+                    {
+                        ""name"" : ""stick"",
+                        ""layout"" : ""Stick"",
+                        ""noisy"" : true
+                    }
+                ]
+            }
+        ";
+
+        InputSystem.RegisterLayout(json);
+        var device = InputSystem.AddDevice("TestLayout");
+
+        Assert.That(device["stick"].As<StickControl>().x.noisy, Is.True);
+        Assert.That(device["stick"].As<StickControl>().y.noisy, Is.True);
+        Assert.That(device["stick"].As<StickControl>().left.noisy, Is.True);
+        Assert.That(device["stick"].As<StickControl>().right.noisy, Is.True);
+        Assert.That(device["stick"].As<StickControl>().up.noisy, Is.True);
+        Assert.That(device["stick"].As<StickControl>().down.noisy, Is.True);
+    }
+
+    [Test]
+    [Category("Layouts")]
     public void Layouts_CanMarkControlAsSynthetic()
     {
         const string json = @"